Add confirmation for deleting an annotation
It's possible for users to assume that the 'x' on the annotator
widget is for closing the window instead of deleting the annotation.

This adds a reminder that they are about to delete their annotation.

+++ src/ui/viewer.js
@@ -1,3 +1,4 @@
+/* global window */
 "use strict";
 
 var Widget = require('./widget').Widget,
@@ -337,11 +338,13 @@ var Viewer = exports.Viewer = Widget.extend({
     //
     // Returns nothing.
     _onDeleteClick: function (event) {
-        var item = $(event.target)
-            .parents('.annotator-annotation')
-            .data('annotation');
-        this.hide();
-        this.options.onDelete(item);
+        if (window.confirm(_t('Delete this annotation?'))) {
+            var item = $(event.target)
+                .parents('.annotator-annotation')
+                .data('annotation');
+            this.hide();
+            this.options.onDelete(item);
+        }
     },
 
     // Event callback: called when a user triggers `mouseover` on a highlight
